Use `ThreadLocal` instead of the tags to store the non-blocking marker (#45)
Since the non-blocking thread marker is immutable, we could store it in a `ThreadLocal`
variable to avoid doing the native code switch.

Given the baseline of 0,495us/op, the overhead of doing a blocking call in non-non-blocking thread is now around 3% (~0,086us/op aka "86 nanoseconds") and comes from the ThreadLocal usage. Which is already a great improvement over 40% overhead of 0,237us/op of the previous version.