Deploy releases to SonatypeOSS (staging repo) instead of Bintray (#182)
This commit switches the target of GA release deployments from Bintray
to Sonatype OSS / Maven Central (triggering the creation of a staging
repository, but not the closure and promotion of said repository).

This will ensure releases can be performed past Bintray's sunset.

Since Maven Central requires signed artifacts, the `signing` plugin is
added to the mix. As a core Gradle plugin, it is trusted enough to have
access to the relevant secrets.

This change also relies on the `release` environment in GitHub having
additional secrets (GPG signing key+passphrase, Sonatype credentials).