Upgrade byte-buddy to version 1.12.21 for support of jdk 18+ versions (#312)
This PR is an attempt to add support for JDK 18+ (18, 19, 20), by upgrading byte-buddy to the latest version which is 1.12.21.

Now, there is an incompatible change that has been made in byte-buddy transform method of the AgentBuilder.Transformer interface, which is implemented by BlockHound. So, the PR adapts to the new signature of the transform method, which is now taking a ProtectionDomainparameter.

Note that you will still need to use the -XX:+AllowRedefinitionToAddDeleteMethods when using JDK13+. Updated the documentation in order to indicate how to add this option for gradle and maven.