Detect jdk major version using Runtime class (#408)
in #407, the BlockHound class is detecting the current jdk major version using InstrumentationUtil.jdkMajorVersion() method that has been introduced in #407 PR (the new InstrumentationUtil.jdkMajorVersion() method is relying on the jdk Runtime.getRuntime() class).

Now, there is still another class (NativeWrappingClassFileTransformer) which needs to determine the current jdk major version. Let's update it in order to reuse the InstrumentationUtil instead of relying of java system properties.